Refactoring Primitive Obsession In the ```Doubleroom``` class
1. Description of the shortcoming
answer :
Indecent Exposure is a code smell that occurs when internal details are unnecessarily exposed. It refers to the situation where methods and variables of a class, which are intended to be used only within that class, are made public or accessible from outside.

The given Doubleroom class, there is an Indecent Exposure issue. The fields name2, contact2, and gender2 are declared as public, which means they can be accessed and modified directly from outside the class.
----------
2. Brief description of the selected refactoring method:

a. What method?
answer :
Encapsulate Field

b. WHY it is chosen?
answer :
One of the core concepts in object-oriented programming is encapsulation, which is the hiding of data within an object. Encapsulating a field allows us to control access to it through getter and setter methods.
----------
3. Conclusion
answer :
Encapsulation is one of the basis of object-oriented programming. By utilising the method of "Encapsulate Field", we construct a clear boundary for accessing and modifying the field values.

Encapsulating fields with getter and setter methods allows us to control the access and modification of the field values. Using getter and setter methods to encapsulate fields can enhance the readability of code by giving descriptive names to the operations of accessing and modifying the field values.

****************************************************************
Reference :
Fowler, M., &amp; Beck, K. (2019a). Encapsulate Field. In Refactoring: Improving the design of existing code (pp. 167–167). essay, Addison-Wesley.

Refactoring.Guru. (n.d.). Encapsulate field. Refactoring.Guru. https://refactoring.guru/encapsulate-field